% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-descriptions.R
\name{add_comp_descriptions}
\alias{add_comp_descriptions}
\title{Add Descriptions to Glycan Compositions}
\usage{
add_comp_descriptions(exp)
}
\arguments{
\item{exp}{An \code{\link[glyexp:experiment]{glyexp::experiment()}} object.}
}
\value{
The experiment object with the new columns added.
}
\description{
This function adds the following columns to the
variable information tibble:
\itemize{
\item 'n_hex': number of Hex
\item 'n_hexnac': number of HexNAc
\item 'n_fuc': number of Fuc
\item 'n_neuac': number of NeuAc
\item 'n_neugc': number of NeuGc
\item 'n_sia': number of Sia
}
}
\details{
\code{n_sia} is the sum of \code{n_neuac} and \code{n_neugc}.

The function relies on the \code{glycan_composition} column in
the \code{var_info} tibble, and it should be a \code{glyrepr_composition} object.
If you use the \code{glyread} package to directly read results from search engines
(such as pGlyco3), this column will be automatically added to \code{var_info}.
If you construct the \code{glyexp_experiment} object yourself, you need to ensure
that this column exists and its type is \code{glyrepr_composition}.
See \code{\link[glyrepr:glycan_composition]{glyrepr::glycan_composition()}} or \code{\link[glyrepr:as_glycan_composition]{glyrepr::as_glycan_composition()}}
for how to create such objects manually.
}
