% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-descriptions.R
\name{add_glycan_descriptions}
\alias{add_glycan_descriptions}
\title{Add Glycan Descriptions}
\usage{
add_glycan_descriptions(exp)
}
\arguments{
\item{exp}{An \code{\link[glyexp:experiment]{glyexp::experiment()}} object.}
}
\value{
The experiment object with the new columns added.
}
\description{
This function adds glycan description columns to the
variable information tibble of an \code{\link[glyexp:experiment]{glyexp::experiment()}} object.
If structure information is available,
both composition and structure descriptions are added.
Otherwise, only composition descriptions are added.
}
\details{
This function is a wrapper around \code{\link[=add_comp_descriptions]{add_comp_descriptions()}}
and \code{\link[=add_struct_descriptions]{add_struct_descriptions()}}.
}
\seealso{
\code{\link[=add_comp_descriptions]{add_comp_descriptions()}}, \code{\link[=add_struct_descriptions]{add_struct_descriptions()}}
}
