% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-descriptions.R
\name{add_struct_descriptions}
\alias{add_struct_descriptions}
\title{Add Descriptions to Glycan Structures}
\usage{
add_struct_descriptions(exp)
}
\arguments{
\item{exp}{An \code{\link[glyexp:experiment]{glyexp::experiment()}} object.}
}
\value{
The experiment object with the new columns added.
}
\description{
This function adds columns about glycan structural properties
to the variable information tibble.
Depending on the glycan type (N-glycan, O-glycan),
different columns are added.
Current, only N-glycan descriptions are implemented.
See \code{\link[=describe_n_glycans]{describe_n_glycans()}} for columns added.
}
\details{
The function relies on the \code{glycan_structure} column in
the \code{var_info} tibble, and it should be a \code{glyrepr_structure} object.
If you use the \code{glyread} package to directly read results from search engines
(such as pGlyco3), this column will be automatically added to \code{var_info}.
If you construct the \code{glyexp_experiment} object yourself, you need to ensure
that this column exists and its type is \code{glyrepr_structure}.
See \code{\link[glyrepr:glycan_structure]{glyrepr::glycan_structure()}} or \code{\link[glyrepr:as_glycan_structure]{glyrepr::as_glycan_structure()}}
for how to create such objects manually.
}
\seealso{
\code{\link[=describe_n_glycans]{describe_n_glycans()}}
}
