% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-motif.R
\name{as.character.glymotif_db_motifs}
\alias{as.character.glymotif_db_motifs}
\title{Coerce Database Motif Vectors to Character}
\usage{
\method{as.character}{glymotif_db_motifs}(x, ...)
}
\arguments{
\item{x}{A \code{glymotif_db_motifs} vector.}

\item{...}{Additional arguments passed to \code{\link[=as.character]{as.character()}}.}
}
\value{
A character vector.
}
\description{
Delegates to the underlying glycan structure character representation.
}
