% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-motif.R
\name{available_motifs}
\alias{available_motifs}
\title{Get Available Motifs}
\usage{
available_motifs()
}
\value{
A character vector of motif names.
}
\description{
This function returns the names of the available motifs available in the package.
}
\examples{
available_motifs()[1:5]

}
