% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts-motif.R
\name{counts_motif}
\alias{counts_motif}
\title{Count How Many Times a Glycan has the Given Motif}
\usage{
counts_motif(glycan, motif, alignment = NULL, ignore_linkages = FALSE)
}
\arguments{
\item{glycan}{A 'glycan_graph' object, or an IUPAC-condensed structure string.}

\item{motif}{A 'glycan_graph' object, an IUPAC-condensed structure string,
or a known motif name (use \code{\link[=available_motifs]{available_motifs()}} to see all available motifs).}

\item{alignment}{A character string.
Possible values are "substructure", "core", "terminal" and "whole".
If not provided, the value will be decided based on the \code{motif} argument.
If \code{motif} is a motif name, the alignment in the database will be used.
Otherwise, "substructure" will be used.}

\item{ignore_linkages}{A logical value. If \code{TRUE}, linkages will be ignored in the comparison.}
}
\value{
An integer indicating how many times the \code{glycan} has the \code{motif}.
}
\description{
This function is closely related to \code{\link[=has_motif]{has_motif()}}.
However, instead of returning a logical value, it returns the number of times
the \code{glycan} has the \code{motif}.
}
\details{
This function actually perform v2f algorithm to get all possible matches
between \code{glycan} and \code{motif}.
However, the result is not necessarily the number of matches.

Think about the following example:
\itemize{
\item glycan: \verb{Gal(b1-?)[Gal(b1-?)]GlcNAc(b1-4)GlcNAc}
\item motif: \verb{Gal(b1-?)[Gal(b1-?)]GlcNAc(b1-)}
}

To draw the glycan out:

\if{html}{\out{<div class="sourceCode">}}\preformatted{Gal 1
   \\ b1-? b1-4
    GlcNAc -- GlcNAc
   / b1-?
Gal 2
}\if{html}{\out{</div>}}

To draw the motif out:

\if{html}{\out{<div class="sourceCode">}}\preformatted{Gal 1
   \\ b1-?
    GlcNAc b1-
   / b1-?
Gal 2
}\if{html}{\out{</div>}}

To differentiate the galactoses, we number them as "Gal 1" and "Gal 2"
in both the glycan and the motif.
The v2f subisomorphic algorithm will return two matches:
\itemize{
\item Gal 1 in the glycan matches Gal 1 in the motif, and Gal 2 matches Gal 2.
\item Gal 1 in the glycan matches Gal 2 in the motif, and Gal 2 matches Gal 1.
}

However, from a biological perspective, the two matches are the same.
This function will take care of this, and return the "unique" number of matches.

For other details about the handling of monosaccharide, linkages, alignment,
substituents, and implementation, see \code{\link[=has_motif]{has_motif()}}.
}
\examples{
counts_motif("Gal(b1-3)Gal(b1-3)GalNAc", "Gal(b1-")
counts_motif(
  "Man(b1-?)[Man(b1-?)]GalNAc(b1-4)GlcNAc",
  "Man(b1-?)[Man(b1-?)]GalNAc"
)
counts_motif("Gal(b1-3)Gal", "Man")

}
\seealso{
\code{\link[=has_motif]{has_motif()}}
}
