% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-motif.R
\name{db_motif_info}
\alias{db_motif_info}
\title{Get Database Motif Information}
\usage{
db_motif_info()
}
\value{
A tibble with \code{source}, \code{source_id}, \code{accession}, \code{name},
\code{alignment}, and \code{glycan_structure} columns.
}
\description{
Returns metadata for all motifs available in the package.
}
\examples{
db_motif_info()

}
