% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-motifs.R
\name{dynamic_motifs}
\alias{dynamic_motifs}
\title{Dynamic Motifs Specification}
\usage{
dynamic_motifs(max_size = 3)
}
\arguments{
\item{max_size}{The maximum number of monosaccharides in the extracted motifs.
Default is 3. Passed to \code{\link[=extract_motif]{extract_motif()}}.}
}
\value{
A \code{dynamic_motifs_spec} object.
}
\description{
Create a specification for dynamic motif extraction.
This should be passed to the \code{motifs} argument of \code{\link[=have_motifs]{have_motifs()}},
\code{\link[=count_motifs]{count_motifs()}}, \code{\link[=match_motifs]{match_motifs()}}, \code{\link[=add_motifs_lgl]{add_motifs_lgl()}}, or \code{\link[=add_motifs_int]{add_motifs_int()}}.
}
\details{
When used, motifs will be extracted dynamically from the input glycans
using \code{\link[=extract_motif]{extract_motif()}}, and motif matching will use \code{alignment = "substructure"}.
}
\examples{
# Use in have_motifs()
# have_motifs(glycans, dynamic_motifs(max_size = 4))

}
\seealso{
\code{\link[=branch_motifs]{branch_motifs()}}, \code{\link[=have_motifs]{have_motifs()}}, \code{\link[=count_motifs]{count_motifs()}}
}
