% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-motif.R
\name{extract_motif}
\alias{extract_motif}
\title{Extract All Substructures (Motifs)}
\usage{
extract_motif(glycans, max_size = 3)
}
\arguments{
\item{glycans}{One of:
\itemize{
\item A \code{\link[glyrepr:glycan_structure]{glyrepr::glycan_structure()}} vector.
\item A glycan structure string vector. All formats supported by \code{\link[glyparse:auto_parse]{glyparse::auto_parse()}} are accepted.
}}

\item{max_size}{The maximum number of monosaccharides in the extracted motifs. Default is 3.
Note that setting this value very large can be computationally expensive.
Try the default value first, and increase it progressively if needed.}
}
\value{
A \code{\link[glyrepr:glycan_structure]{glyrepr::glycan_structure()}} vector containing the unique extracted motifs.
}
\description{
Extract all unique connected subgraphs (motifs) from the input glycans up to a specified size.
This function can be useful combined with \code{count_motifs()} or \code{glydet::quantify_motifs()}.
If so, set \code{alignment} to "substructure" for these functions.

If you want to perform dynamic motif matching in functions like \code{have_motifs()} or \code{glydet::quantify_motifs()},
use the \code{\link[=dynamic_motifs]{dynamic_motifs()}} helper instead,
which handles additional intricacies related to how the motifs should be matched.
}
\examples{
glycan <- "Gal(b1-3)[GlcNAc(a1-6)]GalNAc(a1-"
extract_motif(glycan, max_size = 2)

}
