% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glymotif-package.R
\docType{package}
\name{glymotif-package}
\alias{glymotif}
\alias{glymotif-package}
\title{glymotif: Extract Glycan Motifs from Glycan Structures}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Identify, count, and match recurring substructures in glycan structures. Supports concrete and generic monosaccharide matching, several structural alignment modes, node-to-node mappings, and batch analysis using subgraph isomorphism. Includes curated motif annotations derived from the 'GlycoMotif' resource \url{https://glycomotif.glyomics.org/}. Integrates with 'glyrepr' and 'glyparse' for structural glycomics workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://glycoverse.github.io/glymotif/}
  \item \url{https://github.com/glycoverse/glymotif}
  \item Report bugs at \url{https://github.com/glycoverse/glymotif/issues}
}

}
\author{
\strong{Maintainer}: Bin Fu \email{23110220018@m.fudan.edu.cn} (\href{https://orcid.org/0000-0001-8567-2997}{ORCID}) [copyright holder]

}
\keyword{internal}
