% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-glycan-properties.R
\name{has_bisecting}
\alias{has_bisecting}
\title{Does the Glycan have Bisecting GlcNAc?}
\usage{
has_bisecting(glycan, strict = FALSE)
}
\arguments{
\item{glycan}{A \code{glycan_graph} object, or a character string of IUPAC condensed format.}

\item{strict}{A logical value. If \code{TRUE}, the glycan must have "concrete"
monosaccharides (e.g. "GlcNAc", "Man", "Gal") and linkage information.
If \code{FALSE}, the function is more lenient,
checking monosaacharide identities on the "simple" level (e.g. "H", "N", "F")
and ignoring linkage information.
Default is \code{FALSE}. This is preferred because in most cases the
structural resolution could not be high, but we known for sure the glycans are indeed N-glycans.}
}
\value{
A logical value.
}
\description{
Bisecting GlcNAc is a GlcNAc residue attached to the core mannose of N-glycans.

\if{html}{\out{<div class="sourceCode">}}\preformatted{     Man
        \\
GlcNAc - Man - GlcNAc - GlcNAc
~~~~~~  /
     Man
}\if{html}{\out{</div>}}
}
