% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/have-motif.R
\name{have_motifs}
\alias{have_motifs}
\title{Check if the Glycans have Multiple Motifs}
\usage{
have_motifs(glycans, motifs, alignments = NULL, ignore_linkages = FALSE)
}
\arguments{
\item{glycans}{A 'glyrepr_structure' object, or an IUPAC-condensed structure string vector.}

\item{motifs}{A character vector of motif names, IUPAC-condensed structure strings,
or a list of 'glyrepr_structure' objects.}

\item{alignments}{A character vector specifying alignment types for each motif.
Possible values are "substructure", "core", "terminal" and "whole".
If not provided, the values will be decided based on each motif.
If a motif is a known motif name, the alignment in the database will be used.
Otherwise, "substructure" will be used.
Can be a single value (applied to all motifs) or a vector of the same length as motifs.}

\item{ignore_linkages}{A logical value. If \code{TRUE}, linkages will be ignored in the comparison.}
}
\value{
A tibble where the first column 'glycan' contains glycan identifiers
(names if available, otherwise IUPAC structure strings), and subsequent
columns contain logical values indicating whether each glycan has each motif.
}
\description{
This function checks if the given \code{glycan}s have the given \code{motif}s.
It is a vectorized version of \code{\link[=have_motif]{have_motif()}} that accepts multiple motifs.
The function returns a logical matrix where rows represent glycans and
columns represent motifs.
}
\examples{
library(glyparse)
library(glyrepr)

# Create some glycans
glycan1 <- o_glycan_core_2(mono_type = "concrete")
glycan2 <- parse_iupac_condensed("Gal(b1-?)[GlcNAc(b1-6)]GalNAc")
glycans <- c(glycan1, glycan2)

# Define multiple motifs
motifs <- c("Gal(b1-3)GalNAc", "Gal(b1-4)GalNAc", "GlcNAc(b1-6)GalNAc")

# Check which glycans have which motifs
result <- have_motifs(glycans, motifs)
print(result)

# With different alignment types
alignments <- c("substructure", "substructure", "core")
result2 <- have_motifs(glycans, motifs, alignments = alignments)
print(result2)

}
\seealso{
\code{\link[=have_motif]{have_motif()}}
}
