% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-motif.R
\name{is_known_motif}
\alias{is_known_motif}
\title{Check if a Motif is Known}
\usage{
is_known_motif(name)
}
\arguments{
\item{name}{A character vector of motif names.}
}
\value{
A logical vector.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{is_known_motif()} was deprecated in glymotif 0.16.0. Use
\code{\link[=db_motif_info]{db_motif_info()}} to inspect database motifs instead.
}
\examples{
is_known_motif(c("O-Glycan core 1", "unknown"))
}
