% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-glycan-properties.R
\name{is_n_glycan}
\alias{is_n_glycan}
\title{Check if a Glycan is an N-Glycan}
\usage{
is_n_glycan(glycan, strict = FALSE)
}
\arguments{
\item{glycan}{A \code{glycan_graph} object, or a character string of IUPAC condensed format.}

\item{strict}{A logical value. If \code{TRUE}, the glycan must have "concrete"
monosaccharides (e.g. "GlcNAc", "Man", "Gal") and linkage information.
If \code{FALSE}, the function is more lenient,
checking monosaacharide identities on the "generic" level (e.g. "Hex", "HexNAc")
and ignoring linkage information.
Default is \code{FALSE}. This is preferred because in most cases the
structural resolution could not be high, but we known for sure the glycans are indeed N-glycans.}
}
\value{
A logical value.
}
\description{
This function checks if a glycan has the N-glycan core motif:

\if{html}{\out{<div class="sourceCode">}}\preformatted{Man
   \\
    Man - GlcNAc - GlcNAc
   /
Man
}\if{html}{\out{</div>}}
}
