% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-glycan-properties.R
\name{n_antennae}
\alias{n_antennae}
\title{Number of Antennae}
\usage{
n_antennae(glycan, strict = FALSE)
}
\arguments{
\item{glycan}{A \code{glycan_graph} object, or a character string of IUPAC condensed format.}

\item{strict}{A logical value. If \code{TRUE}, the glycan must have "concrete"
monosaccharides (e.g. "GlcNAc", "Man", "Gal") and linkage information.
If \code{FALSE}, the function is more lenient,
checking monosaacharide identities on the "simple" level (e.g. "H", "N", "F")
and ignoring linkage information.
Default is \code{FALSE}. This is preferred because in most cases the
structural resolution could not be high, but we known for sure the glycans are indeed N-glycans.}
}
\value{
An integer of the number of antennae.
}
\description{
The number of antennae is the number of branching GlcNAc to the core mannoses
in a complex N-glycan. Bisecting GlcNAc is not counted as an antenna.
This functions returns NA_integer_ for non-complex N-glycans.
}
