% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-glycan-properties.R
\name{n_arm_fuc}
\alias{n_arm_fuc}
\title{Number of Arm Focuses}
\usage{
n_arm_fuc(glycan, strict = FALSE)
}
\arguments{
\item{glycan}{A \code{glycan_graph} object, or a character string of IUPAC condensed format.}

\item{strict}{A logical value. If \code{TRUE}, the glycan must have "concrete"
monosaccharides (e.g. "GlcNAc", "Man", "Gal") and linkage information.
If \code{FALSE}, the function is more lenient,
checking monosaacharide identities on the "simple" level (e.g. "H", "N", "F")
and ignoring linkage information.
Default is \code{FALSE}. This is preferred because in most cases the
structural resolution could not be high, but we known for sure the glycans are indeed N-glycans.}
}
\value{
An integer of the number of arm fucoses.
}
\description{
Arm focuses are thoses focuse residues attached to the branching GlcNAc
of an N-glycan.

\if{html}{\out{<div class="sourceCode">}}\preformatted{ Fuc  <- arm fucose
  |
GlcNAc - Man
            \\
             Man - GlcNAc - GlcNAc
            /
GlcNAc - Man
}\if{html}{\out{</div>}}
}
