% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-glycan-properties.R
\name{n_core_fuc}
\alias{n_core_fuc}
\title{Number of Core Fucoses}
\usage{
n_core_fuc(glycan, strict = FALSE)
}
\arguments{
\item{glycan}{A \code{glycan_graph} object, or a character string of IUPAC condensed format.}

\item{strict}{A logical value. If \code{TRUE}, the glycan must have "concrete"
monosaccharides (e.g. "GlcNAc", "Man", "Gal") and linkage information.
If \code{FALSE}, the function is more lenient,
checking monosaacharide identities on the "simple" level (e.g. "H", "N", "F")
and ignoring linkage information.
Default is \code{FALSE}. This is preferred because in most cases the
structural resolution could not be high, but we known for sure the glycans are indeed N-glycans.}
}
\value{
An integer of the number of core fucoses.
}
\description{
Core fucoses are those fucose residues attached to the core GlcNAc of an N-glycan.

\if{html}{\out{<div class="sourceCode">}}\preformatted{Man             Fuc  <- core fucose
   \\             |
    Man - GlcNAc - GlcNAc
   /
Man
}\if{html}{\out{</div>}}
}
