% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-glycan-properties.R
\name{n_terminal_gal}
\alias{n_terminal_gal}
\title{Number of Terminal Galactoses}
\usage{
n_terminal_gal(glycan, strict = FALSE)
}
\arguments{
\item{glycan}{A \code{glycan_graph} object, or a character string of IUPAC condensed format.}

\item{strict}{A logical value. If \code{TRUE}, the glycan must have "concrete"
monosaccharides (e.g. "GlcNAc", "Man", "Gal") and linkage information.
If \code{FALSE}, the function is more lenient,
checking monosaacharide identities on the "simple" level (e.g. "H", "N", "F")
and ignoring linkage information.
Default is \code{FALSE}. This is preferred because in most cases the
structural resolution could not be high, but we known for sure the glycans are indeed N-glycans.}
}
\value{
An integer of the number of terminal galactoses.
}
\description{
Terminal galactoses are those galactose residues on the non-reducing end
without sialic acid capping.

\if{html}{\out{<div class="sourceCode">}}\preformatted{         Gal - GlcNAc - Man
         ~~~               \\
     terminal Gal           GlcNAc - GlcNAc
                           /
Neu5Ac - Gal - GlcNAc - Man
         ~~~
   not terminal Gal
}\if{html}{\out{</div>}}
}
