% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-motifs.R
\name{add_motifs_int}
\alias{add_motifs_int}
\alias{add_motifs_lgl}
\alias{add_motifs_int.glyexp_experiment}
\alias{add_motifs_lgl.glyexp_experiment}
\alias{add_motifs_int.data.frame}
\alias{add_motifs_lgl.data.frame}
\title{Add Motif Annotations}
\usage{
add_motifs_int(x, motifs, alignments = NULL, ignore_linkages = FALSE, ...)

add_motifs_lgl(x, motifs, alignments = NULL, ignore_linkages = FALSE, ...)

\method{add_motifs_int}{glyexp_experiment}(x, motifs, alignments = NULL, ignore_linkages = FALSE, ...)

\method{add_motifs_lgl}{glyexp_experiment}(x, motifs, alignments = NULL, ignore_linkages = FALSE, ...)

\method{add_motifs_int}{data.frame}(x, motifs, alignments = NULL, ignore_linkages = FALSE, ...)

\method{add_motifs_lgl}{data.frame}(x, motifs, alignments = NULL, ignore_linkages = FALSE, ...)
}
\arguments{
\item{x}{A \code{\link[glyexp:experiment]{glyexp::experiment()}} object, or a tibble with a structure column.}

\item{motifs}{A character vector of motif names, glycan structure strings,
or a 'glyrepr_structure' object.
For glycan structure strings, all formats supported by \code{\link[glyparse:auto_parse]{glyparse::auto_parse()}} are accepted,
including IUPAC-condensed, WURCS, GlycoCT, and others.}

\item{alignments}{A character vector specifying alignment types for each motif.
Can be a single value (applied to all motifs) or a vector of the same length as motifs.}

\item{ignore_linkages}{A logical value. If \code{TRUE}, linkages will be ignored in the comparison.}

\item{...}{Additional arguments passed to the method.}
}
\value{
An \code{\link[glyexp:experiment]{glyexp::experiment()}} object with motif annotations added to the variable information.
}
\description{
This function adds motif annotations to the variable information
of a \code{\link[glyexp:experiment]{glyexp::experiment()}} or a tibble with a structure column.
\code{add_motifs_int()} adds integer annotations (how many motifs are present).
\code{add_motifs_lgl()} adds boolean annotations (whether the motif is present).
}
\section{About Names}{


The naming rule for the new columns is similar to that of \code{\link[=have_motifs]{have_motifs()}}.
Briefly, you can use named character vector to name the motifs,
and that will be used as the new column names.
The only catchup is that you cannot pass a named \code{glyrepr::glycan_structure()} to \code{motifs}.
This is a fundamental limitation of the \code{vctrs_rcrd} class,
which \code{glyrepr::glycan_structure()} is built on.
}

\section{Why do we need these functions}{


Adding one motif annotation to a \code{\link[glyexp:experiment]{glyexp::experiment()}} is easy:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{exp |>
  mutate_var(has_hex = have_motif(glycan_structure, "Hex"))
}\if{html}{\out{</div>}}

However, adding multiple motifs is not as straightforward.
You can still use \code{mutate_var()} to add multiple motifs like this:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{exp |>
  mutate_var(
    n_hex = count_motif(glycan_structure, "Hex"),
    n_dhex = count_motif(glycan_structure, "dHex"),
    n_hexnac = count_motif(glycan_structure, "HexNAc"),
  )
}\if{html}{\out{</div>}}

This method has two problems:
\enumerate{
\item it has a lot of boilerplate code (a lot of typing)
\item it is not very efficient, as each call to \code{count_motif}
performs validation and conversion on \code{glycan_structure},
which is a time-consuming process.
}

Therefore, we think it would be better to have a function that
adds multiple motif annotations in a single call, in a more intuitive way.
That's why we provide these two functions.

Under the hood, they use a more straightforward approach for \code{\link[glyexp:experiment]{glyexp::experiment()}} objects:
\enumerate{
\item get the motif annotation matrix using \code{count_motifs()} or \code{have_motifs()}
\item convert the matrix to a tibble
\item use \code{dplyr::bind_cols()} to add the tibble to the variable information
}
}

\seealso{
\code{\link[=have_motifs]{have_motifs()}}, \code{\link[=count_motifs]{count_motifs()}}, \code{\link[glyexp:experiment]{glyexp::experiment()}}
}
