% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-motifs.R
\name{branch_motifs}
\alias{branch_motifs}
\title{Branch Motifs Specification}
\usage{
branch_motifs()
}
\value{
A \code{branch_motifs_spec} object.
}
\description{
Create a specification for branch motif extraction.
This should be passed to the \code{motifs} argument of \code{\link[=have_motifs]{have_motifs()}},
\code{\link[=count_motifs]{count_motifs()}}, \code{\link[=match_motifs]{match_motifs()}}, \code{\link[=add_motifs_lgl]{add_motifs_lgl()}}, or \code{\link[=add_motifs_int]{add_motifs_int()}}.
}
\details{
When used, motifs will be extracted dynamically from the input glycans
using \code{\link[=extract_branch_motif]{extract_branch_motif()}} with \code{including_core = TRUE}.
Motif matching will use a custom \code{match_degree} where the last 4 nodes
of each motif are not required to match degree exactly (to allow for
attachment to the glycan core).
}
\examples{
# Use in have_motifs()
# have_motifs(glycans, branch_motifs())

}
\seealso{
\code{\link[=dynamic_motifs]{dynamic_motifs()}}, \code{\link[=have_motifs]{have_motifs()}}, \code{\link[=count_motifs]{count_motifs()}}
}
