% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-motif.R
\name{db_motif_info}
\alias{db_motif_info}
\title{Get Database Motif Information}
\usage{
db_motif_info()
}
\value{
A tibble.
}
\description{
Returns metadata for all motifs available in the package.
You can use \code{dplyr::distinct(db_motif_info(), source_id, source)} to get
all available sources.
}
\details{
It contains the following columns:
\itemize{
\item \code{source_id}: the collection identifier of the motif
\item \code{source}: the collection name of the motif
\item \code{accession}: the accession number of the motif
\item \code{name}: the name of the motif
\item \code{alignment}: the alignment of the motif
\item \code{glycan_structure}: the glycan structure (glyrepr::glycan_structure()) of the motif
}
}
\examples{
db_motif_info()

}
