% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-motif.R
\name{db_motifs}
\alias{db_motifs}
\title{Get All Motifs from the Database}
\usage{
db_motifs()
}
\value{
A \code{\link[glyrepr:glycan_structure]{glyrepr::glycan_structure()}} vector.
}
\description{
This function returns all motifs available in the package as a glycan
structure vector.
We use GlycoMotif GlyGen Collection (https://glycomotif.glyomics.org/glycomotif/GGM) as the source of the motifs.
This function is useful to be integrated with \code{\link[=have_motifs]{have_motifs()}} and \code{\link[=count_motifs]{count_motifs()}}.
For example, use \code{have_motifs(glycans, db_motifs())} to check against all motifs.
}
\details{
The returned vector carries database \code{alignment} and \code{name} metadata so
motif-matching functions can use GlycoMotif alignment rules directly.
}
\examples{
db_motifs()[1:5]

}
