% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-motif.R
\name{db_motifs}
\alias{db_motifs}
\title{Get All Motifs from the Database}
\usage{
db_motifs()
}
\value{
A \code{db_motifs_spec} object.
}
\description{
This function returns a database motif specification.
We use GlycoMotif GlyGen Collection (https://glycomotif.glyomics.org/glycomotif/GGM) as the source of the motifs.
This function is useful to be integrated with \code{\link[=have_motifs]{have_motifs()}} and \code{\link[=count_motifs]{count_motifs()}}.
For example, use \code{have_motifs(glycans, db_motifs())} to check against all motifs.
}
\details{
Use \code{\link[=db_motif_info]{db_motif_info()}} to inspect the motifs included in the database.
}
\examples{
db_motifs()
db_motif_info()

}
