% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-motif.R
\name{extract_branch_motif}
\alias{extract_branch_motif}
\title{Extract Branch Motifs}
\usage{
extract_branch_motif(glycans)
}
\arguments{
\item{glycans}{One of:
\itemize{
\item A \code{\link[glyrepr:glycan_structure]{glyrepr::glycan_structure()}} vector.
\item A glycan structure string vector. All formats supported by \code{\link[glyparse:auto_parse]{glyparse::auto_parse()}} are accepted.
}}
}
\value{
A \code{\link[glyrepr:glycan_structure]{glyrepr::glycan_structure()}} vector containing the unique extracted branching motifs.
}
\description{
An N-glycan branching motif if the substructures linked to either the a3- or a6-core-mannose.

For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{Neu5Ac - Gal - GlcNAc - Man
~~~~~~~~~~~~~~~~~~~~~      \\
  A branching motif         Man - GlcNAc - GlcNAc -
                           /
                        Man
}\if{html}{\out{</div>}}

This function returns all the unique branching motifs found in the input glycans.
It is very useful for analyzing N-glycan antennary patterns combined with \code{glydet::quantify_motifs()}.
When used with \code{have_motif()} or related functions,
we recommend setting \code{alignment} to "substructure".
}
\details{
The function works by:
\enumerate{
\item Converting the input to a set of unique \code{glycan_structure} objects.
\item Searching for the N-glycan branch pattern: \verb{HexNAc(??-?)Hex(??-?)Hex(??-?)HexNAc(??-?)HexNAc(??-}.
\item For each match, identifying the root node of the branch (the leftmost HexNAc in the pattern).
\item Extracting the full subtree rooted at that node.
\item Preserving the correct anomeric configuration (e.g., "b1") by inspecting the linkage to the root node.
}
}
\examples{
glycans <- c(
  "Neu5Ac(a2-3)Gal(b1-4)GlcNAc(b1-2)Man(a1-3)[Man(a1-6)]Man(b1-4)GlcNAc(a1-4)GlcNAc(b1-",
  "Gal(b1-4)GlcNAc(b1-2)Man(a1-3)[Man(a1-6)]Man(b1-4)GlcNAc(a1-4)GlcNAc(b1-"
)
extract_branch_motif(glycans)

}
