% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-motif.R
\name{get_motif_graph}
\alias{get_motif_graph}
\alias{get_motif_alignment}
\alias{get_motif_aglycon}
\title{Get a Motif Structure Graph, Alignment, or Aglycon}
\usage{
get_motif_graph(name)

get_motif_alignment(name)

get_motif_aglycon(name)
}
\arguments{
\item{name}{A character string of the motif name.}
}
\value{
\code{get_motif_graph()} returns a \code{glycan_graph} object,
\code{get_motif_alignment()} returns a character string, and
\code{get_motif_aglycon()} returns a character string.
}
\description{
Given a motif name in the GlycoMotif GlyGen Collection,
\code{get_motif_graph()} returns a \code{glycan_graph} object of the motif structure,
\code{get_motif_alignment()} returns the alignment of the motif, and
\code{get_motif_aglycon()} returns the aglycon of the motif.
For available motifs, use \code{\link[=availabel_motifs]{availabel_motifs()}}.
}
\examples{
get_motif_graph("N-Glycan core basic")
get_motif_alignment("N-Glycan core basic")
get_motif_aglycon("N-Glycan core basic")

}
\seealso{
\code{\link[=availabel_motifs]{availabel_motifs()}}, \code{\link[=is_known_motif]{is_known_motif()}}
}
