% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-motif.R
\name{get_motif_structure}
\alias{get_motif_structure}
\alias{get_motif_alignment}
\title{Get the Structures or Alignments of Known Motifs}
\usage{
get_motif_structure(name)

get_motif_alignment(name)
}
\arguments{
\item{name}{A character vector of motif names.}
}
\value{
\itemize{
\item \code{get_motif_structure()}: a \code{\link[glyrepr:glycan_structure]{glyrepr::glycan_structure()}} vector.
\item \code{get_motif_alignment()}: a character vector of motif alignments.
}

For \code{get_motif_alignment()}, if \code{name} has length greater than 1, the return
value is named with the motif names.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{get_motif_structure()} and \code{get_motif_alignment()} were deprecated in
glymotif 0.16.0. Use \code{\link[=db_motif_info]{db_motif_info()}} to inspect database motifs instead.
}
\examples{
get_motif_structure("N-Glycan core basic")
get_motif_alignment("N-Glycan core basic")

get_motif_structure(c("O-Glycan core 1", "O-Glycan core 2"))
get_motif_alignment(c("O-Glycan core 1", "O-Glycan core 2"))

}
\seealso{
\code{\link[=db_motif_info]{db_motif_info()}}
}
