% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-motif.R
\name{get_motif_structure}
\alias{get_motif_structure}
\alias{get_motif_alignment}
\alias{get_motif_aglycon}
\title{Get the Structures, Alignments, or Aglycons of Known Motifs}
\usage{
get_motif_structure(name)

get_motif_alignment(name)

get_motif_aglycon(name)
}
\arguments{
\item{name}{A character vector of the motif name.}
}
\value{
\code{get_motif_structure()} returns a \code{glycan_graph} when \code{name} is a character
scalar, and a list of \code{glycan_graph} when \code{name} is a character vector.
\code{get_motif_alignment()} returns a character vector of motif alignments.
\code{get_motif_aglycon()} returns a character vector of motif aglycons.
For all three functions, if \code{name} has length greater than 1,
the return value is named with the motif names.
}
\description{
Given a character vector of motifs names in GlycoMotif GlyGen Collection,
these functions return the structures, alignments, or aglycons of the motifs.
}
\examples{
get_motif_structure("N-Glycan core basic")
get_motif_alignment("N-Glycan core basic")
get_motif_aglycon("N-Glycan core basic")

get_motif_structure(c("O-Glycan core 1", "O-Glycan core 2"))
get_motif_alignment(c("O-Glycan core 1", "O-Glycan core 2"))
get_motif_aglycon(c("O-Glycan core 1", "O-Glycan core 2"))

}
\seealso{
\code{\link[=available_motifs]{available_motifs()}}, \code{\link[=is_known_motif]{is_known_motif()}}
}
