% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glymotif-package.R
\docType{package}
\name{glymotif-package}
\alias{glymotif}
\alias{glymotif-package}
\title{glymotif: Extract glycan motifs from glycan structures}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package extracts glycan motifs (substructures) from glycan structures. It also determine common structural properties of N-glycans, e.g. the number of antennae, the number of branches, etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fubin1999.github.io/glymotif/}
}

}
\author{
\strong{Maintainer}: Bin Fu \email{23110220018@m.fudan.edu.cn} (\href{https://orcid.org/0000-0001-8567-2997}{ORCID})

}
\keyword{internal}
