% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has-motif-vectorized.R
\name{has_motifs}
\alias{has_motifs}
\title{Check if a Glycan have the Given Motifs}
\usage{
has_motifs(
  glycan,
  motifs = NULL,
  ...,
  alignments = "substructure",
  ignore_linkages = FALSE
)
}
\arguments{
\item{glycan}{A 'glycan_graph' object, or an IUPAC-condensed structure string.}

\item{motifs}{A list of 'glycan_graph' objects, a character vector of
IUPAC-condensed structure strings, or a character vector of known motif names.}

\item{...}{Ignored.}

\item{alignments}{A character vector of alignment names for the motifs,
either of the same length as \code{motifs}, or a single character string.
Could be "substructure", "core", "terminal", or "whole".
Default to "substructure".
If not provided and \code{motifs} are known motif names or missing (using all known motifs),
the alignments in the GlycoMotif GlyGen Collection will be used.}

\item{ignore_linkages}{A logical value. If \code{TRUE}, linkages will be ignored in the comparison.}
}
\value{
A logical vector, indicating if the glycan has the motifs.
}
\description{
This is a vectorized versions of \code{\link[=has_motif]{has_motif()}}.
It checks if a glycan has the given motifs, or all known motifs
in the GlycoMotif GlyGen Collection if \code{motifs} is missing.

A logical vector is returned.
Names of \code{motifs} are preserved as names of the output logical vectors.
If no names are provided, and the input is a character vector, the names
will be the vector itself.
}
\examples{
library(glyparse)

# Check if a glycan has the given motifs using names.
# Note that the `motifs` vector will become names of the result vector.
glycan <- "Gal(b1-3)GlcNAc(b1-3)Gal(b1-3)GalNAc(a1-"
has_motifs(glycan, c("O-Glycan core 1", "O-Glycan core 2"))

# Using IUPAC-condensed structure strings.
# Note that names of `motifs` are preserved in the result vector.
motifs <- c(M1 = "Gal(b1-3)GalNAc(a1-", M2 = "GlcNAc(b1-4)GlcNAc(a1-")
has_motifs(glycan, motifs)

# Or using 'glycan_graph' objects.
motifs <- list(
  M1 = parse_iupac_condensed("Gal(b1-3)GalNAc(a1-"),
  M2 = parse_iupac_condensed("GlcNAc(b1-4)GlcNAc(a1-")
)
has_motifs(glycan, motifs)

# If motifs are missing, use all known motifs in the
# GlycoMotif GlyGen Collection.
has_motifs(glycan)[1:5]

}
