% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has-motif-vectorized.R
\name{has_motifs}
\alias{has_motifs}
\alias{have_motif}
\title{Check if Glycan(s) have the Given Motif(s)}
\usage{
has_motifs(glycan, motifs = NULL, ...)

have_motif(glycans, motif, ...)
}
\arguments{
\item{glycan}{A 'glycan_graph' object, or an IUPAC-condensed structure string.}

\item{motifs}{A list of 'glycan_graph' objects, a character vector of
IUPAC-condensed structure strings, or a character vector of known motif names.}

\item{...}{Additional arguments passed to \code{\link[=has_motif]{has_motif()}}.}

\item{glycans}{A list of 'glycan_graph' objects, or a character vector of
IUPAC-condensed structure strings.}

\item{motif}{A 'glycan_graph' object, an IUPAC-condensed structure string,
or a known motif name (use \code{\link[=available_motifs]{available_motifs()}} to see all available motifs).}
}
\value{
A logical vector, indicating if the glycan(s) have the motif(s).
}
\description{
These are vectorized versions of \code{\link[=has_motif]{has_motif()}}, vectorizing over
motifs (\code{has_motifs()}) and glycans (\code{have_motif()}), respectively.
They can be told apart by English grammar :).

Names of glycans or motifs
are preserved as names of the output logical vectors.
If no names are provided, and the input is a character vector, the names
will be the vector itself.
It is always recommended to provide names for the input vectors.

If \code{motifs} are missing in \code{has_motifs()}, all known motifs from
GlycoMotif GlyGen Collection will be used.
}
\examples{
library(glyparse)

# Check if a glycan has the given motifs using names
glycan <- "Gal(b1-3)GlcNAc(b1-3)Gal(b1-3)GalNAc(a1-"
has_motifs(glycan, c("O-Glycan core 1", "O-Glycan core 2"))

# using IUPAC-condensed structure strings
has_motifs(glycan, c("Gal(b1-3)GalNAc(a1-", "GlcNAc(b1-4)GlcNAc(a1-"))

# or using 'glycan_graph' objects
motifs <- list(
  M1 = parse_iupac_condensed("Gal(b1-3)GalNAc(a1-"),
  M2 = parse_iupac_condensed("GlcNAc(b1-4)GlcNAc(a1-")
)
has_motifs(glycan, motifs)

# If motifs are missing, use all known motifs
has_motifs(glycan)[1:5]

# Check if glycans have the given motif
glycans <- c(
  G1 = "Gal(b1-3)GlcNAc(b1-3)Gal(b1-3)Gal",
  G2 = "Gal(b1-3)GlcNAc(b1-3)Gal(b1-3)GalNAc(a1-"
)
have_motif(glycans, "Gal(b1-3)Gal")

}
