% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has-motif-vectorized.R
\name{has_motifs}
\alias{has_motifs}
\alias{have_motif}
\title{Check if Glycan(s) have the Given Motif(s)}
\usage{
has_motifs(glycan, motifs, ...)

have_motif(glycans, motif, ...)
}
\arguments{
\item{glycan}{A 'glycan_graph' object, or an IUPAC-condensed structure string.}

\item{motifs}{A list of 'glycan_graph' objects, a character vector of
IUPAC-condensed structure strings, or a character vector of known motif names.}

\item{...}{Additional arguments passed to \code{\link[=has_motif]{has_motif()}}.}

\item{glycans}{A list of 'glycan_graph' objects, or a character vector of
IUPAC-condensed structure strings.}

\item{motif}{A 'glycan_graph' object, an IUPAC-condensed structure string,
or a known motif name (use \code{\link[=available_motifs]{available_motifs()}} to see all available motifs).}
}
\description{
These are vectorized versions of \code{\link[=has_motif]{has_motif()}}, vectorizing over
motifs (\code{has_motifs}) and glycans (\code{have_motif}), respectively.
They can be told apart by English grammar :).
}
