% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has-motif-vectorized.R
\name{have_motif}
\alias{have_motif}
\title{Check if these Glycans have the Given Motif}
\usage{
have_motif(
  glycans,
  motif,
  ...,
  alignment = "substructure",
  ignore_linkages = FALSE
)
}
\arguments{
\item{glycans}{A list of 'glycan_graph' objects, or a character vector of IUPAC-condensed structure strings.}

\item{motif}{A 'glycan_graph' object, an IUPAC-condensed structure string, or a known motif name.}

\item{...}{Ignored.}

\item{alignment}{A character string of alignment name for the motif.
Could be "substructure", "core", "terminal", or "whole".
Default to "substructure".
If not provided and \code{motif} is a known motif name,
the alignment in the GlycoMotif GlyGen Collection will be used.}

\item{ignore_linkages}{A logical value. If \code{TRUE}, linkages will be ignored in the comparison.}
}
\value{
A logical vector, indicating if the glycans have the motif.
}
\description{
This is a vectorized version of \code{\link[=has_motif]{has_motif()}}.
It checks if a list of glycans (or a character vector of IUPAC-condensed structure strings)
have the given motif.

A logical vector is returned.
Names of \code{glycans} are preserved as names of the output logical vectors.
If no names are provided, and the input is a character vector, the names
will be the vector itself.
}
\examples{
glycan <- c(G1 = "Gal(b1-3)GlcNAc", G2 = "Man(b1-4)GlcNAc", G3 = "GlcNAc")
have_motif(glycan, "Gal")

}
