% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-motif-graph.R
\name{is_known_motif}
\alias{is_known_motif}
\title{Check if a Motif is Known}
\usage{
is_known_motif(name)
}
\arguments{
\item{name}{A character string of the motif name.}
}
\value{
A logical value.
}
\description{
This function checks if a motif is a known motif in GlycoMotif GlyGen Collection.
}
