% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-motif.R
\name{is_known_motif}
\alias{is_known_motif}
\title{Check if a Motif is Known}
\usage{
is_known_motif(name)
}
\arguments{
\item{name}{A character vector of the motif name.}
}
\value{
A logical vector.
}
\description{
This function checks if motifs are known motifs in GlycoMotif GlyGen Collection.
}
