% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-glycan-properties.R
\name{n_glycan_type}
\alias{n_glycan_type}
\title{Determine N-Glycan Type}
\usage{
n_glycan_type(glycan, strict = FALSE)
}
\arguments{
\item{glycan}{A \code{glycan_graph} object, or a character string of IUPAC condensed format.}

\item{strict}{A logical value. If \code{TRUE}, the glycan must have "concrete"
monosaccharides (e.g. "GlcNAc", "Man", "Gal") and linkage information.
If \code{FALSE}, the function is more lenient,
checking monosaacharide identities on the "simple" level (e.g. "H", "N", "F")
and ignoring linkage information.
Default is \code{FALSE}. This is preferred because in most cases the
structural resolution could not be high, but we known for sure the glycans are indeed N-glycans.}
}
\value{
A character string of the N-glycan type,
either "highmannose", "hybrid", "complex", or "paucimannose".
If the glycan seems not to be any of the four types, an error is thrown
with a message "Not an N-glycan".
This doesn't necessarily mean the glycan is not an N-glycan.
Maybe you have used the strict mode with a glycan that is not well resolved.
}
\description{
Four types of N-glycans are recognized: high mannose, hybrid, complex, and paucimannose.
For more information about N-glycan types,
see \href{https://www.ncbi.nlm.nih.gov/books/NBK579964/#_s9_2_}{Essentials of Glycobiology}.
}
