% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify-motifs.R
\name{quantify_motifs}
\alias{quantify_motifs}
\title{Quantify motifs in an experiment}
\usage{
quantify_motifs(exp, motifs, alignments = NULL, ignore_linkages = FALSE)
}
\arguments{
\item{exp}{A \code{glyexp::experiment()} object containing glycoproteomics data.}

\item{motifs}{A character vector of motif names, glycan structure strings,
or a 'glyrepr_structure' object.
For glycan structure strings, all formats supported by \code{\link[glyparse:auto_parse]{glyparse::auto_parse()}} are accepted,
including IUPAC-condensed, WURCS, GlycoCT, and others.}

\item{alignments}{A character vector specifying alignment types for each motif.
Can be a single value (applied to all motifs) or a vector of the same length as motifs.}

\item{ignore_linkages}{A logical value. If \code{TRUE}, linkages will be ignored in the comparison.}
}
\value{
A \code{glyexp::experiment()} object containing motif quantifications.
}
\description{
This function takes a \code{glyexp::experiment()},
and returns a new \code{glyexp::experiment()} with motif quantifications.

The new experiment is different from a normal \code{glyexp::experiment()} in three ways:
\enumerate{
\item It doesn't have a "glycan_structure" or "glycan_composition" column,
but a "motif" column instead.
\item The metadata field "exp_type" is "motifomics", not "glycoproteomics" or glycomics.
}

You can understand a "motif experiment" in this way:
For glycoproteomics data,
instead of containing the quantification of each glycan on each glycosite,
it now contains the quantification of each motif on each glycosite in each sample.
For glycomics data,
it contains the motif quantification in each sample.

Due to the unified data structure of \code{glyexp::experiment()},
the returned motif experiment can be passed to downstream \code{glycoverse}
functions like \code{glystats::gly_ttest()} for further statistical analysis.
Also, you can use \code{as_tibble()} to convert it to a "tidy" tibble for custom analysis.
}
\details{
Here is a comprehensive comparison of the input experiment and the returned experiment:

\strong{Sample information:}
Sample information doesn't change.

\strong{Variable information:}
Firstly, we define the "base variables" as the unique combination of all columns
in the variable information tibble except for "variable", "glycan_composition", and "glycan_structure".
The new var_info would be a Cartesian product of all "base variables" and all motifs.
For glycomics data, no "base variable" is defined,
so the resulted var_info only contains the "motif" column.

\strong{Expression matrix:}
Rows are variables, columns are samples, same as before.
Filling the cells are the quantifications for motifs on different glycosites in each sample,
instead of quantifications of glycoforms or glycopeptides.
Or just motif quantification in each sample, for glycomics data.
}
\seealso{
\code{\link[=count_motifs]{count_motifs()}}
}
